requirements.txt
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
binance-connector==1.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kaleido==0.2.1
lxml==4.8.0
multitasking==0.0.10
numpy==1.22.2
pandas==1.4.1
plotly==5.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scipy==1.8.0
service-identity==21.1.0
six==1.16.0
ta==0.9.0
tenacity==8.0.1
Twisted==22.1.0
txaio==21.2.1
typing-extensions==4.1.1
urllib3==1.26.8
yfinance==0.1.70
zope.interface==5.4.0
